lib/notification_service/requirements.txt
#
# Copyright 2022 The AI Flow Authors
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#   http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing,
# software distributed under the License is distributed on an
# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
# KIND, either express or implied.  See the License for the
# specific language governing permissions and limitations
# under the License.
#

protobuf==3.15.6
grpcio==1.35.0
sqlalchemy>=1.3.18, <2
pyyaml>=5.1, <5.4
rich==9.2.0
python-dateutil==2.8.0
alembic>=1.2, <2.0
python-daemon>=2.1.1, <2.2
